Rename efficient_capsnet to src
Initially, I intended to keep the Efficient-CapsNet code seperate from my own work. However, I will likely have to edit quite some things in the original code (changing how the dataset it loaded, changing the capsules themselves for example). Therefor, it makes more sense to treat this directory as the src directory. An additional benefit is that this prevents annoying imports as well. Also, changed the original README (the one belonging to Efficient-CapsNet) to reflect this.